+++ docs/src/news.md
@@ -3,7 +3,7 @@
 ## *Version 0.10.2 dev*
 
 - News
-    - Function **occmwpca**: One-class classification (OCC) by moving window Pca.
+    - Function **occmwpca**: One-class classification (OCC) by moving window Pca (MWPCA).
 
 - Modifications
     - Functions **outsd**, **outod**, **outsdod**, **occsd**, **occod**, **occsdod**,